doc: fix in amb macro.
* txr.1: when the amb macro detects that the continuation has
succeeded, it should return that successful value from the
amb-scope, rather than returning the local successful argument
a from the amb function. Although it works both ways, it is
inefficient when it returns from the function. The reason is
that each call to amb executes the successful future twice:
once via (call cont a) and then again by returning the a
value. This then causes an exponential cascade in calls: each
successive amb call sthe successful future twice, so for
instance if the solution has a sequence of 8 amb calls, the
successful case is reached 256 times.