vim: improvement multi-line strings.
Vim's handling of multi-line Lisp strings is glitchy. We are
contributing to it by tryign to match the backslash-newline as
an escape sequence.

As a result of this change, Vim is less confused.  Indentation
is still incorrect after some multi-line strings, but I'm not
seeing the discrepancy between the behavior of the visual
parentheses matching, and the % parentheses jump.
I'm able to navigate around in the stdlib/getopts.tl code.

* genvim.txr (chesc): Remove the backslash-newline from the
list of character escapes.
(txr_string, txr_quasilit, txr_regex, tx_regex): Use skip= to
recognize the backslash-newline sequence as part of the
literal.