tree: fix array underruns found by ubsan.
* tree.c (tn_find_low): In this function it may happen that
we are looking for a low bound that is beyond the highest value in the
tree. In this situation, tdi->lastnode is nil; the loop that walks back
up the path will not find it, and underrun. Now the word before
the path array is the tree_iter_state, which is zero. That zero word
terminates the loop so things work by fluke.
(tr_find_rebuild_scapegoat): Firstly, since in this function we
pre-decrement ti->depth, we guard the whole thing by ti->depth > 0.
This is just a precaution; the algorithm guarantees that a scapegoat
will be found, so this will not recurse to the depth == 0 case.
The main fix here is this:  we sometimes call tr_rebuild in the
case when the parent is the root node. In that case, there is no
grandparent. The path is empty and we are accessing before the array
to try to fetch the nonexistent grandparent. Again, by fluke,
there is a zero there thanks to the tree_iter_state which turns
into a null pointer that tells tr_rebuild that the node has no parent.
The correct way is to check the depth and pass nil as the grandparent
if ti->depth is zero.