compiler: autoload on sys:*in-compilation-unit*.
The compiler module contains the with-compilation-unit macro;
if that macro is used, it is auto-loaded. However, when a file
that uses with-compilation-unit is compiled, the macro is
gone; its expansion refers to the sys:*in-compilation-unit*
which is defined in the compiler module also. However, there
is no autoload set up for it.

* lisplib.c (compiler_set_entries): Add
sys:*in-compilation-unit* to the list of symbols that trigger
autoloading.