configure: bugfix in shell escaping code
* configure: in the case when the argument contains
single quotes, we try to identify the subcase that cannot use
double quotes instead. That subcase is not correctly
identified: it occurs when there are double quotes, dollar
signs **or backslashes**. For instance the data '\ cannot
be put into double quotes as "'\" on account of the backslash,
which is an active character in double quotes. We also
simplify the pattern by using a character class.