New FFI types str-s, bstr-s and wstr-s.
These types actually make it possible to receive a string by
pointer from a C function, without trying to free it.
It is now possible to write a FFI wrapper for strtol or
wcstol, which is done in the new test case.

* ffi.c (str_s_s, bstr_s_s, wstr_s_s): New symbol variables.
(ffi_init_types): Register the types str-s, bstr-s and wstr-s.
(ffi_init): Intern the new symbols.

* tests/017/str-s.tl: New file.

* txr.1: Documented.

* stdlib/doc-syms.tl: Updated.