fork/streams: new failing test.
This test case exemplifies code that will work as expected
when *stdout* is a TTY device, such that line buffering is
in effect, but then break when standard output is redirected
to a file.

The issue is that the controlling process is not flushing its
standard output when calling the external script, so the
script's output gets placed ahead of the process' own earlier
output.

* tests/018/forkflush.tl: New file.

* tests/018/forkflush.expected: New file.