Support RISC-V: conditionally.
Everything compiles, and most tests pass. Failing test are:

- tests/014/dgram-stream.tl:
- tests/017/glob-carray.tl:
- tests/017/qsort.tl:

The datagram test is not able to bind a UDP socket to a port.
On the gcc402 machine of the GCC Compile Farm, attempts bind a
UDP socket to any of the ports 1024-65535 all fail with errno
99 (EADDRNOTAVAIL).

The FFI test cases are segfaulting. The qosrt.tl case has
the problem in the test cases which abort the callback with
a return-from.

* unwind.h (struct jmp): Define for RISC-V.

* jmp.S (DEFUN): Define macro for RISC-V.
(jmp_save, jmp_restore): Define for RISC-V.